+++ hive_generator/example/lib/hive_registrar.g.dart
@@ -2,11 +2,11 @@ import 'package:hive_ce/hive.dart';
 import 'package:example/types.dart';
 
 extension HiveRegistrar on HiveInterface {
-  static void registerAdapters() {
-    Hive.registerAdapter(Class1Adapter());
-    Hive.registerAdapter(Class2Adapter());
-    Hive.registerAdapter(EmptyClassAdapter());
-    Hive.registerAdapter(IterableClassAdapter());
-    Hive.registerAdapter(Enum1Adapter());
+  void registerAdapters() {
+    registerAdapter(Class1Adapter());
+    registerAdapter(Class2Adapter());
+    registerAdapter(EmptyClassAdapter());
+    registerAdapter(IterableClassAdapter());
+    registerAdapter(Enum1Adapter());
   }
 }